+++ app/src/main/java/com/teslasoft/jarvis/Executor.java
@@ -1,9 +1,0 @@
-package com.teslasoft.jarvis;
-
-public class Executor
-{
-	public int Execute(String name, String args[])
-	{
-		return 0;
-	}
-}