+++ app/src/main/java/com/teslasoft/jarvis/Launcher.java
@@ -148,7 +148,7 @@ public class Launcher extends Activity
 	@Override
 	public void onBackPressed()
 	{
-		// super.onBackPressed();
+		super.onBackPressed();
 	}
 	
 	public void GetDate ()