+++ app/src/main/java/com/teslasoft/libraries/support/AdminReceiver.java
@@ -4,39 +4,22 @@ import android.os.Build;
 import android.content.Intent;
 import android.content.Context;
 import android.content.ComponentName;
-import android.widget.Toast;
 import android.app.admin.DeviceAdminReceiver;
-import android.app.admin.DevicePolicyManager;
+
+// Antiviruses can mark this file as Android/LockScreen.Jisut.AX
 
 public class AdminReceiver extends DeviceAdminReceiver
 {
 	public void onEnabled(Context context, Intent intent) 
 	{
-		Toast.makeText(context.getApplicationContext(), "Device Admin enabled", Toast.LENGTH_LONG).show();
-		context.startService(new Intent(context.getApplicationContext(), CheckinService.class));
+		
 	}
 	
-	public CharSequence onDisableRequested(Context context, Intent intent)
-	{
-		return "Warning: Some services will not work";
-	}
-
 	public void onDisabled(Context context, Intent intent)
 	{
-		Toast.makeText(context.getApplicationContext(), "Device Admin disabled", Toast.LENGTH_LONG).show();
-		context.stopService(new Intent(context.getApplicationContext(), CheckinService.class));
-	}
-
-	public void onLockTaskModeEntering(Context context, Intent intent,String pkg)
-	{
 		
 	}
-
-	public void onLockTaskModeExiting(Context context, Intent intent)
-	{
-		
-	}
-
+	
 	public static ComponentName getComponentName(Context context)
 	{
 		return new ComponentName(context.getApplicationContext(), AdminReceiver.class);