+++ app/src/main/java/io/jarvis/In.java
@@ -1,6 +1,0 @@
-package io.jarvis;
-
-public class In
-{
-	
-}